+++ src/cloudfront/lambda-request-handler/package.json
@@ -16,5 +16,8 @@
   "bugs": {
     "url": "https://github.com/FusionAuth/fusionauth-site/issues"
   },
-  "homepage": "https://github.com/FusionAuth/fusionauth-site#readme"
+  "homepage": "https://github.com/FusionAuth/fusionauth-site#readme",
+  "dependencies": {
+    "aws-sdk": "^2.1480.0"
+  }
 }
