--- src/cloudfront/lambda-request-handler/src/fusionauth-website-lambda-request-handler.js
@@ -1,619 +1,0 @@
-const referenceData = {
-  redirects: {
-    '/docs/v1/tech/apis/connectors/generic': '/docs/apis/connectors/generic',
-    '/docs/v1/tech/apis/connectors/ldap': '/docs/apis/connectors/ldap',
-    '/docs/v1/tech/apis/connectors/': '/docs/apis/connectors/',
-    '/docs/v1/tech/apis/form-fields': '/docs/apis/custom-forms/form-fields',
-    '/docs/v1/tech/apis/forms': '/docs/apis/custom-forms/forms',
-    '/docs/v1/tech/apis/entity-management/entities': '/docs/apis/entities/entities',
-    '/docs/v1/tech/apis/entity-management/entity-types': '/docs/apis/entities/entity-types',
-    '/docs/v1/tech/apis/entity-management/grants': '/docs/apis/entities/grants',
-    '/docs/v1/tech/apis/entity-management/': '/docs/apis/entities/',
-    '/docs/v1/tech/apis/identity-providers/apple': '/docs/apis/identity-providers/apple',
-    '/docs/v1/tech/apis/identity-providers/epicgames': '/docs/apis/identity-providers/epicgames',
-    '/docs/v1/tech/apis/identity-providers/external-jwt': '/docs/apis/identity-providers/external-jwt',
-    '/docs/v1/tech/apis/identity-providers/facebook': '/docs/apis/identity-providers/facebook',
-    '/docs/v1/tech/apis/identity-providers/google': '/docs/apis/identity-providers/google',
-    '/docs/v1/tech/apis/identity-providers/hypr': '/docs/apis/identity-providers/hypr',
-    '/docs/v1/tech/apis/identity-providers/linkedin': '/docs/apis/identity-providers/linkedin',
-    '/docs/v1/tech/apis/identity-providers/links': '/docs/apis/identity-providers/links',
-    '/docs/v1/tech/apis/identity-providers/nintendo': '/docs/apis/identity-providers/nintendo',
-    '/docs/v1/tech/apis/identity-providers/openid-connect': '/docs/apis/identity-providers/openid-connect',
-    '/docs/v1/tech/apis/identity-providers/': '/docs/apis/identity-providers/',
-    '/docs/v1/tech/apis/identity-providers/samlv2': '/docs/apis/identity-providers/samlv2',
-    '/docs/v1/tech/apis/identity-providers/samlv2-idp-initiated': '/docs/apis/identity-providers/samlv2-idp-initiated',
-    '/docs/v1/tech/apis/identity-providers/sonypsn': '/docs/apis/identity-providers/sonypsn',
-    '/docs/v1/tech/apis/identity-providers/steam': '/docs/apis/identity-providers/steam',
-    '/docs/v1/tech/apis/identity-providers/twitch': '/docs/apis/identity-providers/twitch',
-    '/docs/v1/tech/apis/identity-providers/twitter': '/docs/apis/identity-providers/twitter',
-    '/docs/v1/tech/apis/identity-providers/xbox': '/docs/apis/identity-providers/xbox',
-    '/docs/v1/tech/apis/messengers/generic': '/docs/apis/messengers/generic',
-    '/docs/v1/tech/apis/two-factor': '/docs/apis/messengers/two-factor',
-    '/docs/v1/tech/apis/messengers/': '/docs/apis/messengers/',
-    '/docs/v1/tech/apis/messengers/twilio': '/docs/apis/messengers/twilio',
-    '/docs/v1/tech/apis/scim/': '/docs/apis/scim/',
-    '/docs/v1/tech/apis/scim/SCIMEnterpriseUser': '/docs/apis/scim/SCIMEnterpriseUser',
-    '/docs/v1/tech/apis/scim/SCIMGroup': '/docs/apis/scim/SCIMGroup',
-    '/docs/v1/tech/apis/scim/SCIMServiceProvider': '/docs/apis/scim/SCIMServiceProvider',
-    '/docs/v1/tech/apis/scim/SCIMUser': '/docs/apis/scim/SCIMUser',
-    '/docs/v1/tech/apis/actioning-users': '/docs/apis/actioning-users',
-    '/docs/v1/tech/apis/api-explorer': '/docs/apis/api-explorer',
-    '/docs/v1/tech/apis/api-keys': '/docs/apis/api-keys',
-    '/docs/v1/tech/apis/applications': '/docs/apis/applications',
-    '/docs/v1/tech/apis/audit-logs': '/docs/apis/audit-logs',
-    '/docs/v1/tech/apis/authentication': '/docs/apis/authentication',
-    '/docs/v1/tech/apis/consents': '/docs/apis/consents',
-    '/docs/v1/tech/apis/emails': '/docs/apis/emails',
-    '/docs/v1/tech/apis/errors': '/docs/apis/errors',
-    '/docs/v1/tech/apis/event-logs': '/docs/apis/event-logs',
-    '/docs/v1/tech/apis/families': '/docs/apis/families',
-    '/docs/v1/tech/apis/groups': '/docs/apis/groups',
-    '/docs/v1/tech/apis/hosted-backend': '/docs/apis/hosted-backend',
-    '/docs/v1/tech/apis/integrations': '/docs/apis/integrations',
-    '/docs/v1/tech/apis/ip-acl': '/docs/apis/ip-acl',
-    '/docs/v1/tech/apis/jwt': '/docs/apis/jwt',
-    '/docs/v1/tech/apis/keys': '/docs/apis/keys',
-    '/docs/v1/tech/apis/lambdas': '/docs/apis/lambdas',
-    '/docs/v1/tech/apis/login': '/docs/apis/login',
-    '/docs/v1/tech/apis/message-templates': '/docs/apis/message-templates',
-    '/docs/v1/tech/apis/': '/docs/apis/',
-    '/docs/v1/tech/apis/passwordless': '/docs/apis/passwordless',
-    '/docs/v1/tech/apis/reactor': '/docs/apis/reactor',
-    '/docs/v1/tech/apis/registrations': '/docs/apis/registrations',
-    '/docs/v1/tech/apis/reports': '/docs/apis/reports',
-    '/docs/v1/tech/apis/system': '/docs/apis/system',
-    '/docs/v1/tech/apis/tenants': '/docs/apis/tenants',
-    '/docs/v1/tech/apis/themes': '/docs/apis/themes',
-    '/docs/v1/tech/apis/user-action-reasons': '/docs/apis/user-action-reasons',
-    '/docs/v1/tech/apis/user-actions': '/docs/apis/user-actions',
-    '/docs/v1/tech/apis/user-comments': '/docs/apis/user-comments',
-    '/docs/v1/tech/apis/users': '/docs/apis/users',
-    '/docs/v1/tech/apis/webauthn': '/docs/apis/webauthn',
-    '/docs/v1/tech/apis/webhooks': '/docs/apis/webhooks',
-    '/docs/v1/tech/integrations/twilio': '/docs/customize/email-and-messages/deprecated/twilio',
-    '/docs/v1/tech/tutorials/two-factor/': '/docs/customize/email-and-messages/deprecated/',
-    '/docs/v1/tech/tutorials/two-factor/twilio-push-pre-1-26': '/docs/customize/email-and-messages/deprecated/twilio-push-pre-1-26',
-    '/docs/v1/tech/tutorials/two-factor/authenticator-app-pre-1-26': '/docs/customize/email-and-messages/deprecated/authenticator-app-pre-1-26',
-    '/docs/v1/tech/guides/configuring-application-specific-email-templates': '/docs/customize/email-and-messages/configuring-application-specific-email-templates',
-    '/docs/v1/tech/email-templates/configure-email': '/docs/customize/email-and-messages/configure-email',
-    '/docs/v1/tech/email-templates/email-templates': '/docs/customize/email-and-messages/email-templates',
-    '/docs/v1/tech/email-templates/templates-replacement-variables': '/docs/customize/email-and-messages/templates-replacement-variables',
-    '/docs/v1/tech/messengers/generic-messenger': '/docs/customize/email-and-messages/generic-messenger',
-    '/docs/v1/tech/email-templates/message-templates': '/docs/customize/email-and-messages/message-templates',
-    '/docs/v1/tech/messengers/twilio-messenger': '/docs/customize/email-and-messages/twilio-messenger',
-    '/docs/v1/tech/email-templates/': '/docs/customize/email-and-messages/',
-    '/docs/v1/tech/messengers/': '/docs/customize/email-and-messages/',
-    '/docs/v1/tech/themes/application-specific-themes': '/docs/customize/look-and-feel/application-specific-themes',
-    '/docs/v1/tech/admin-guide/cli': '/docs/customize/look-and-feel/cli',
-    '/docs/v1/tech/themes/client-side-password-rule-validation': '/docs/customize/look-and-feel/client-side-password-rule-validation',
-    '/docs/v1/tech/themes/examples': '/docs/customize/look-and-feel/examples',
-    '/docs/v1/tech/themes/helpers': '/docs/customize/look-and-feel/helpers',
-    '/docs/v1/tech/themes/kickstart-custom-theme': '/docs/customize/look-and-feel/kickstart-custom-theme',
-    '/docs/v1/tech/themes/localization': '/docs/customize/look-and-feel/localization',
-    '/docs/v1/tech/themes/tailwind': '/docs/customize/look-and-feel/tailwind',
-    '/docs/v1/tech/themes/template-variables': '/docs/customize/look-and-feel/template-variables',
-    '/docs/v1/tech/themes/': '/docs/customize/look-and-feel/',
-    '/docs/v1/tech/lambdas/': '/docs/extend/code/lambdas/',
-    '/docs/v1/tech/plugins/custom-password-hashing': '/docs/extend/code/password-hashes/custom-password-hashing',
-    '/docs/v1/tech/plugins/writing-a-plugin': '/docs/extend/code/password-hashes/writing-a-plugin',
-    '/docs/v1/tech/plugins/': '/docs/extend/code/password-hashes/',
-    '/docs/v1/tech/events-webhooks/events/audit-log-create': '/docs/extend/events-and-webhooks/events/audit-log-create',
-    '/docs/v1/tech/events-webhooks/events/event-log-create': '/docs/extend/events-and-webhooks/events/event-log-create',
-    '/docs/v1/tech/events-webhooks/events/group-create': '/docs/extend/events-and-webhooks/events/group-create',
-    '/docs/v1/tech/events-webhooks/events/group-create-complete': '/docs/extend/events-and-webhooks/events/group-create-complete',
-    '/docs/v1/tech/events-webhooks/events/group-delete': '/docs/extend/events-and-webhooks/events/group-delete',
-    '/docs/v1/tech/events-webhooks/events/group-delete-complete': '/docs/extend/events-and-webhooks/events/group-delete-complete',
-    '/docs/v1/tech/events-webhooks/events/group-member-add': '/docs/extend/events-and-webhooks/events/group-member-add',
-    '/docs/v1/tech/events-webhooks/events/group-member-add-complete': '/docs/extend/events-and-webhooks/events/group-member-add-complete',
-    '/docs/v1/tech/events-webhooks/events/group-member-remove': '/docs/extend/events-and-webhooks/events/group-member-remove',
-    '/docs/v1/tech/events-webhooks/events/group-member-remove-complete': '/docs/extend/events-and-webhooks/events/group-member-remove-complete',
-    '/docs/v1/tech/events-webhooks/events/group-member-update': '/docs/extend/events-and-webhooks/events/group-member-update',
-    '/docs/v1/tech/events-webhooks/events/group-member-update-complete': '/docs/extend/events-and-webhooks/events/group-member-update-complete',
-    '/docs/v1/tech/events-webhooks/events/group-update': '/docs/extend/events-and-webhooks/events/group-update',
-    '/docs/v1/tech/events-webhooks/events/group-update-complete': '/docs/extend/events-and-webhooks/events/group-update-complete',
-    '/docs/v1/tech/events-webhooks/events/jwt-public-key-update': '/docs/extend/events-and-webhooks/events/jwt-public-key-update',
-    '/docs/v1/tech/events-webhooks/events/jwt-refresh': '/docs/extend/events-and-webhooks/events/jwt-refresh',
-    '/docs/v1/tech/events-webhooks/events/jwt-refresh-token-revoke': '/docs/extend/events-and-webhooks/events/jwt-refresh-token-revoke',
-    '/docs/v1/tech/events-webhooks/events/kickstart-success': '/docs/extend/events-and-webhooks/events/kickstart-success',
-    '/docs/v1/tech/events-webhooks/events/user-two-factor-method-add': '/docs/extend/events-and-webhooks/events/user-two-factor-method-add',
-    '/docs/v1/tech/events-webhooks/events/user-two-factor-method-remove': '/docs/extend/events-and-webhooks/events/user-two-factor-method-remove',
-    '/docs/v1/tech/events-webhooks/events/user-actions': '/docs/extend/events-and-webhooks/events/user-actions',
-    '/docs/v1/tech/events-webhooks/events/user-bulk-create': '/docs/extend/events-and-webhooks/events/user-bulk-create',
-    '/docs/v1/tech/events-webhooks/events/user-create': '/docs/extend/events-and-webhooks/events/user-create',
-    '/docs/v1/tech/events-webhooks/events/user-create-complete': '/docs/extend/events-and-webhooks/events/user-create-complete',
-    '/docs/v1/tech/events-webhooks/events/user-deactivate': '/docs/extend/events-and-webhooks/events/user-deactivate',
-    '/docs/v1/tech/events-webhooks/events/user-delete': '/docs/extend/events-and-webhooks/events/user-delete',
-    '/docs/v1/tech/events-webhooks/events/user-delete-complete': '/docs/extend/events-and-webhooks/events/user-delete-complete',
-    '/docs/v1/tech/events-webhooks/events/user-email-update': '/docs/extend/events-and-webhooks/events/user-email-update',
-    '/docs/v1/tech/events-webhooks/events/user-email-verified': '/docs/extend/events-and-webhooks/events/user-email-verified',
-    '/docs/v1/tech/events-webhooks/events/user-identity-provider-link': '/docs/extend/events-and-webhooks/events/user-identity-provider-link',
-    '/docs/v1/tech/events-webhooks/events/user-identity-provider-unlink': '/docs/extend/events-and-webhooks/events/user-identity-provider-unlink',
-    '/docs/v1/tech/events-webhooks/events/user-login-failed': '/docs/extend/events-and-webhooks/events/user-login-failed',
-    '/docs/v1/tech/events-webhooks/events/user-login-id-duplicate-create': '/docs/extend/events-and-webhooks/events/user-login-id-duplicate-create',
-    '/docs/v1/tech/events-webhooks/events/user-login-id-duplicate-update': '/docs/extend/events-and-webhooks/events/user-login-id-duplicate-update',
-    '/docs/v1/tech/events-webhooks/events/user-login-new-device': '/docs/extend/events-and-webhooks/events/user-login-new-device',
-    '/docs/v1/tech/events-webhooks/events/user-login-success': '/docs/extend/events-and-webhooks/events/user-login-success',
-    '/docs/v1/tech/events-webhooks/events/user-login-suspicious': '/docs/extend/events-and-webhooks/events/user-login-suspicious',
-    '/docs/v1/tech/events-webhooks/events/user-password-breach': '/docs/extend/events-and-webhooks/events/user-password-breach',
-    '/docs/v1/tech/events-webhooks/events/user-password-reset-send': '/docs/extend/events-and-webhooks/events/user-password-reset-send',
-    '/docs/v1/tech/events-webhooks/events/user-password-reset-start': '/docs/extend/events-and-webhooks/events/user-password-reset-start',
-    '/docs/v1/tech/events-webhooks/events/user-password-reset-success': '/docs/extend/events-and-webhooks/events/user-password-reset-success',
-    '/docs/v1/tech/events-webhooks/events/user-password-update': '/docs/extend/events-and-webhooks/events/user-password-update',
-    '/docs/v1/tech/events-webhooks/events/user-reactivate': '/docs/extend/events-and-webhooks/events/user-reactivate',
-    '/docs/v1/tech/events-webhooks/events/user-registration-create': '/docs/extend/events-and-webhooks/events/user-registration-create',
-    '/docs/v1/tech/events-webhooks/events/user-registration-create-complete': '/docs/extend/events-and-webhooks/events/user-registration-create-complete',
-    '/docs/v1/tech/events-webhooks/events/user-registration-delete': '/docs/extend/events-and-webhooks/events/user-registration-delete',
-    '/docs/v1/tech/events-webhooks/events/user-registration-delete-complete': '/docs/extend/events-and-webhooks/events/user-registration-delete-complete',
-    '/docs/v1/tech/events-webhooks/events/user-registration-update': '/docs/extend/events-and-webhooks/events/user-registration-update',
-    '/docs/v1/tech/events-webhooks/events/user-registration-update-complete': '/docs/extend/events-and-webhooks/events/user-registration-update-complete',
-    '/docs/v1/tech/events-webhooks/events/user-registration-verified': '/docs/extend/events-and-webhooks/events/user-registration-verified',
-    '/docs/v1/tech/events-webhooks/events/user-update': '/docs/extend/events-and-webhooks/events/user-update',
-    '/docs/v1/tech/events-webhooks/events/user-update-complete': '/docs/extend/events-and-webhooks/events/user-update-complete',
-    '/docs/v1/tech/integrations/kafka': '/docs/extend/events-and-webhooks/kafka',
-    '/docs/v1/tech/events-webhooks/': '/docs/extend/events-and-webhooks/',
-    '/docs/v1/tech/events-webhooks/writing-a-webhook': '/docs/extend/events-and-webhooks/writing-a-webhook',
-    '/docs/v1/tech/events-webhooks/securing': '/docs/extend/events-and-webhooks/securing',
-    '/docs/v1/tech/developer-guide/api-gateways/aws-api-gateway': '/docs/extend/examples/api-gateways/aws-api-gateway',
-    '/docs/v1/tech/developer-guide/api-gateways/haproxy-api-gateway': '/docs/extend/examples/api-gateways/haproxy-api-gateway',
-    '/docs/v1/tech/developer-guide/api-gateways/kong-gateway': '/docs/extend/examples/api-gateways/kong-gateway',
-    '/docs/v1/tech/developer-guide/api-gateways/ngrok-cloud-edge': '/docs/extend/examples/api-gateways/ngrok-cloud-edge',
-    '/docs/v1/tech/guides/api-authorization': '/docs/extend/examples/api-gateways/api-authorization',
-    '/docs/v1/tech/example-apps/': '/docs/extend/examples/example-apps/',
-    '/docs/v1/tech/guides/multi-tenant': '/docs/extend/examples/multi-tenant',
-    '/docs/v1/tech/core-concepts/applications': '/docs/get-started/core-concepts/applications',
-    '/docs/v1/tech/core-concepts/authentication-authorization': '/docs/get-started/core-concepts/authentication-authorization',
-    '/docs/v1/tech/core-concepts/editions-features': '/docs/get-started/core-concepts/editions-features',
-    '/docs/v1/tech/core-concepts/entity-management': '/docs/get-started/core-concepts/entity-management',
-    '/docs/v1/tech/core-concepts/groups': '/docs/get-started/core-concepts/groups',
-    '/docs/v1/tech/core-concepts/identity-providers': '/docs/get-started/core-concepts/identity-providers',
-    '/docs/v1/tech/core-concepts/integration-points': '/docs/get-started/core-concepts/integration-points',
-    '/docs/v1/tech/reference/limitations': '/docs/get-started/core-concepts/limitations',
-    '/docs/v1/tech/admin-guide/licensing': '/docs/get-started/core-concepts/licensing',
-    '/docs/v1/tech/core-concepts/localization-and-internationalization': '/docs/get-started/core-concepts/localization-and-internationalization',
-    '/docs/v1/tech/core-concepts/registrations': '/docs/get-started/core-concepts/registrations',
-    '/docs/v1/tech/core-concepts/roles': '/docs/get-started/core-concepts/roles',
-    '/docs/v1/tech/core-concepts/tenants': '/docs/get-started/core-concepts/tenants',
-    '/docs/v1/tech/core-concepts/users': '/docs/get-started/core-concepts/users',
-    '/docs/v1/tech/developer-guide/exposing-instance': '/docs/get-started/download-and-install/development/exposing-instance',
-    '/docs/v1/tech/installation-guide/kickstart': '/docs/get-started/download-and-install/development/kickstart',
-    '/docs/v1/tech/installation-guide/kubernetes/setup/eks': '/docs/get-started/download-and-install/kubernetes/eks',
-    '/docs/v1/tech/installation-guide/kubernetes/fusionauth-deployment': '/docs/get-started/download-and-install/kubernetes/fusionauth-deployment',
-    '/docs/v1/tech/installation-guide/kubernetes/setup/gke': '/docs/get-started/download-and-install/kubernetes/gke',
-    '/docs/v1/tech/installation-guide/kubernetes/setup/aks': '/docs/get-started/download-and-install/kubernetes/aks',
-    '/docs/v1/tech/installation-guide/kubernetes/setup/minikube': '/docs/get-started/download-and-install/kubernetes/minikube',
-    '/docs/v1/tech/admin-guide/account-portal': '/docs/get-started/download-and-install/account-portal',
-    '/docs/v1/tech/installation-guide/common-configuration': '/docs/get-started/download-and-install/common-configuration',
-    '/docs/v1/tech/installation-guide/database': '/docs/get-started/download-and-install/database',
-    '/docs/v1/tech/installation-guide/docker': '/docs/get-started/download-and-install/docker',
-    '/docs/v1/tech/installation-guide/fast-path': '/docs/get-started/download-and-install/fast-path',
-    '/docs/v1/tech/installation-guide/fusionauth-app': '/docs/get-started/download-and-install/fusionauth-app',
-    '/docs/v1/tech/installation-guide/fusionauth-search': '/docs/get-started/download-and-install/fusionauth-search',
-    '/docs/v1/tech/installation-guide/homebrew': '/docs/get-started/download-and-install/homebrew',
-    '/docs/v1/tech/installation-guide/packages': '/docs/get-started/download-and-install/packages',
-    '/docs/v1/tech/installation-guide/server-layout': '/docs/get-started/download-and-install/server-layout',
-    '/docs/v1/tech/tutorials/setup-wizard': '/docs/get-started/download-and-install/setup-wizard',
-    '/docs/v1/tech/guides/silent-mode': '/docs/get-started/download-and-install/silent-mode',
-    '/docs/v1/tech/tutorials/start-and-stop': '/docs/get-started/download-and-install/start-and-stop',
-    '/docs/v1/tech/installation-guide/system-requirements': '/docs/get-started/download-and-install/system-requirements',
-    '/docs/v1/tech/installation-guide/cloud': '/docs/get-started/run-in-the-cloud/cloud',
-    '/docs/v1/tech/installation-guide/marketplaces': '/docs/get-started/run-in-the-cloud/marketplaces',
-    '/docs/v1/tech/tutorials/switch-search-engines': '/docs/lifecycle/manage-users/search/switch-search-engines',
-    '/docs/v1/tech/identity-providers/samlv2/adfs': '/docs/lifecycle/authenticate-users/identity-providers/enterprise/adfs',
-    '/docs/v1/tech/identity-providers/openid-connect/azure-ad': '/docs/lifecycle/authenticate-users/identity-providers/enterprise/azure-ad',
-    '/docs/v1/tech/identity-providers/samlv2/azure-ad': '/docs/lifecycle/authenticate-users/identity-providers/enterprise/azure-ad',
-    '/docs/v1/tech/identity-providers/hypr': '/docs/lifecycle/authenticate-users/identity-providers/enterprise/hypr',
-    '/docs/v1/tech/identity-providers/openid-connect/okta': '/docs/lifecycle/authenticate-users/identity-providers/enterprise/okta',
-    '/docs/v1/tech/identity-providers/samlv2-idp-initiated/okta': '/docs/lifecycle/authenticate-users/identity-providers/enterprise/okta',
-    '/docs/v1/tech/identity-providers/external-jwt/example': '/docs/lifecycle/authenticate-users/identity-providers/external-jwt/example',
-    '/docs/v1/tech/identity-providers/external-jwt/': '/docs/lifecycle/authenticate-users/identity-providers/external-jwt/',
-    '/docs/v1/tech/identity-providers/openid-connect/discord': '/docs/lifecycle/authenticate-users/identity-providers/gaming/discord',
-    '/docs/v1/tech/identity-providers/epic-games': '/docs/lifecycle/authenticate-users/identity-providers/gaming/epic-games',
-    '/docs/v1/tech/identity-providers/nintendo': '/docs/lifecycle/authenticate-users/identity-providers/gaming/nintendo',
-    '/docs/v1/tech/identity-providers/sony': '/docs/lifecycle/authenticate-users/identity-providers/gaming/sony',
-    '/docs/v1/tech/identity-providers/steam': '/docs/lifecycle/authenticate-users/identity-providers/gaming/steam',
-    '/docs/v1/tech/identity-providers/twitch': '/docs/lifecycle/authenticate-users/identity-providers/gaming/twitch',
-    '/docs/v1/tech/identity-providers/xbox': '/docs/lifecycle/authenticate-users/identity-providers/gaming/xbox',
-    '/docs/v1/tech/identity-providers/openid-connect/cognito': '/docs/lifecycle/authenticate-users/identity-providers/social/cognito',
-    '/docs/v1/tech/identity-providers/apple': '/docs/lifecycle/authenticate-users/identity-providers/social/apple',
-    '/docs/v1/tech/identity-providers/facebook': '/docs/lifecycle/authenticate-users/identity-providers/social/facebook',
-    '/docs/v1/tech/identity-providers/openid-connect/github': '/docs/lifecycle/authenticate-users/identity-providers/social/github',
-    '/docs/v1/tech/identity-providers/google': '/docs/lifecycle/authenticate-users/identity-providers/social/google',
-    '/docs/v1/tech/identity-providers/linkedin': '/docs/lifecycle/authenticate-users/identity-providers/social/linkedin',
-    '/docs/v1/tech/identity-providers/twitter': '/docs/lifecycle/authenticate-users/identity-providers/social/twitter',
-    '/docs/v1/tech/identity-providers/openid-connect/': '/docs/lifecycle/authenticate-users/identity-providers/',
-    '/docs/v1/tech/identity-providers/samlv2/': '/docs/lifecycle/authenticate-users/identity-providers/',
-    '/docs/v1/tech/oauth/integrations/cockroachdb': '/docs/lifecycle/authenticate-users/integrations/oidc/cockroachdb',
-    '/docs/v1/tech/oauth/integrations/salesforce': '/docs/lifecycle/authenticate-users/integrations/oidc/salesforce',
-    '/docs/v1/tech/samlv2/aiven': '/docs/lifecycle/authenticate-users/integrations/saml/aiven',
-    '/docs/v1/tech/samlv2/google': '/docs/lifecycle/authenticate-users/integrations/saml/google',
-    '/docs/v1/tech/samlv2/pagerduty': '/docs/lifecycle/authenticate-users/integrations/saml/pagerduty',
-    '/docs/v1/tech/samlv2/sendgrid': '/docs/lifecycle/authenticate-users/integrations/saml/sendgrid',
-    '/docs/v1/tech/samlv2/tableau-cloud': '/docs/lifecycle/authenticate-users/integrations/saml/tableau-cloud',
-    '/docs/v1/tech/samlv2/zendesk': '/docs/lifecycle/authenticate-users/integrations/saml/zendesk',
-    '/docs/v1/tech/tutorials/json-web-tokens': '/docs/lifecycle/authenticate-users/login-api/json-web-tokens',
-    '/docs/v1/tech/oauth/endpoints': '/docs/lifecycle/authenticate-users/oauth/endpoints',
-    '/docs/v1/tech/core-concepts/modes': '/docs/lifecycle/authenticate-users/oauth/modes',
-    '/docs/v1/tech/oauth/tokens': '/docs/lifecycle/authenticate-users/oauth/tokens',
-    '/docs/v1/tech/oauth/url-validation': '/docs/lifecycle/authenticate-users/oauth/url-validation',
-    '/docs/v1/tech/oauth/': '/docs/lifecycle/authenticate-users/oauth/',
-    '/docs/v1/tech/passwordless/magic-links': '/docs/lifecycle/authenticate-users/passwordless/magic-links',
-    '/docs/v1/tech/admin-guide/webauthn': '/docs/lifecycle/authenticate-users/passwordless/webauthn',
-    '/docs/v1/tech/passwordless/webauthn-passkeys': '/docs/lifecycle/authenticate-users/passwordless/webauthn-passkeys',
-    '/docs/v1/tech/passwordless/': '/docs/lifecycle/authenticate-users/passwordless/',
-    '/docs/v1/tech/samlv2/': '/docs/lifecycle/authenticate-users/saml/',
-    '/docs/v1/tech/tutorials/application-authentication-tokens': '/docs/lifecycle/authenticate-users/application-authentication-tokens',
-    '/docs/v1/tech/guides/multi-factor-authentication': '/docs/lifecycle/authenticate-users/multi-factor-authentication',
-    '/docs/v1/tech/guides/single-sign-on': '/docs/lifecycle/authenticate-users/single-sign-on',
-    '/docs/v1/tech/tutorials/gating/setting-up-user-account-lockout': '/docs/lifecycle/authenticate-users/setting-up-user-account-lockout',
-    '/docs/v1/tech/account-management/two-factor-authenticator': '/docs/lifecycle/manage-users/account-management/two-factor-authenticator',
-    '/docs/v1/tech/account-management/two-factor-email': '/docs/lifecycle/manage-users/account-management/two-factor-email',
-    '/docs/v1/tech/account-management/two-factor-sms': '/docs/lifecycle/manage-users/account-management/two-factor-sms',
-    '/docs/v1/tech/account-management/add-webauthn': '/docs/lifecycle/manage-users/account-management/add-webauthn',
-    '/docs/v1/tech/account-management/bootstrapping-login': '/docs/lifecycle/manage-users/account-management/bootstrapping-login',
-    '/docs/v1/tech/account-management/customizing-account-management': '/docs/lifecycle/manage-users/account-management/customizing-account-management',
-    '/docs/v1/tech/account-management/troubleshooting': '/docs/lifecycle/manage-users/account-management/troubleshooting',
-    '/docs/v1/tech/account-management/updating-user-data': '/docs/lifecycle/manage-users/account-management/updating-user-data',
-    '/docs/v1/tech/core-concepts/search': '/docs/lifecycle/manage-users/search/search',
-    '/docs/v1/tech/guides/user-search-with-elasticsearch': '/docs/lifecycle/manage-users/search/user-search-with-elasticsearch',
-    '/docs/v1/tech/tutorials/gating/registration-gate-accounts-until-verified': '/docs/lifecycle/manage-users/verification/registration-gate-accounts-until-verified',
-    '/docs/v1/tech/tutorials/gating/gate-accounts-until-user-email-verified': '/docs/lifecycle/manage-users/verification/gate-accounts-until-user-email-verified',
-    '/docs/v1/tech/guides/registration-email-verification': '/docs/lifecycle/manage-users/verification/registration-email-verification',
-    '/docs/v1/tech/integrations/cleanspeak': '/docs/lifecycle/manage-users/cleanspeak',
-    '/docs/v1/tech/guides/user-actions': '/docs/lifecycle/manage-users/user-actions',
-    '/docs/v1/tech/migration-guide/cognito': '/docs/lifecycle/migrate-users/bulk/cognito',
-    '/docs/v1/tech/migration-guide/auth0': '/docs/lifecycle/migrate-users/bulk/auth0',
-    '/docs/v1/tech/guides/auth0-migration': '/docs/lifecycle/migrate-users/bulk/auth0',
-    '/docs/v1/tech/migration-guide/duende': '/docs/lifecycle/migrate-users/bulk/duende',
-    '/docs/v1/tech/migration-guide/firebase': '/docs/lifecycle/migrate-users/bulk/firebase',
-    '/docs/v1/tech/migration-guide/general': '/docs/lifecycle/migrate-users/bulk/general',
-    '/docs/v1/tech/migration-guide/keycloak': '/docs/lifecycle/migrate-users/bulk/keycloak',
-    '/docs/v1/tech/migration-guide/azureadb2c': '/docs/lifecycle/migrate-users/bulk/azureadb2c',
-    '/docs/v1/tech/migration-guide/supabase': '/docs/lifecycle/migrate-users/bulk/supabase',
-    '/docs/v1/tech/migration-guide/tutorial': '/docs/lifecycle/migrate-users/bulk/tutorial',
-    '/docs/v1/tech/connectors/fusionauth-connector': '/docs/lifecycle/migrate-users/connectors/fusionauth-connector',
-    '/docs/v1/tech/connectors/generic-connector': '/docs/lifecycle/migrate-users/connectors/generic-connector',
-    '/docs/v1/tech/connectors/ldap-connector': '/docs/lifecycle/migrate-users/connectors/ldap-connector',
-    '/docs/v1/tech/connectors/': '/docs/lifecycle/migrate-users/connectors/',
-    '/docs/v1/tech/premium-features/scim/azure-ad-client': '/docs/lifecycle/migrate-users/scim/azure-ad-client',
-    '/docs/v1/tech/premium-features/scim/okta-client': '/docs/lifecycle/migrate-users/scim/okta-client',
-    '/docs/v1/tech/core-concepts/scim': '/docs/lifecycle/migrate-users/scim/scim',
-    '/docs/v1/tech/premium-features/scim/scim-sdk': '/docs/lifecycle/migrate-users/scim/scim-sdk',
-    '/docs/v1/tech/guides/advanced-registration-forms': '/docs/lifecycle/register-users/advanced-registration-forms',
-    '/docs/v1/tech/guides/anonymous-user': '/docs/lifecycle/register-users/anonymous-user',
-    '/docs/v1/tech/tutorials/register-user-login-api': '/docs/lifecycle/register-users/register-user-login-api',
-    '/docs/v1/tech/guides/basic-registration-forms': '/docs/lifecycle/register-users/basic-registration-forms',
-    '/docs/v1/tech/installation-guide/cluster': '/docs/operate/deploy/cluster',
-    '/docs/v1/tech/admin-guide/configuration-management': '/docs/operate/deploy/configuration-management',
-    '/docs/v1/tech/admin-guide/proxy-setup': '/docs/operate/deploy/proxy-setup',
-    '/docs/v1/tech/admin-guide/upgrade': '/docs/operate/deploy/upgrade',
-    '/docs/v1/tech/admin-guide/user-support-guide': '/docs/operate/deploy/user-support-guide',
-    'example-scripts': '/docs/v1/tech/docs/operate/deploy/example-scripts',
-    '/docs/v1/tech/admin-guide/releases': '/docs/operate/roadmap/releases',
-    '/docs/v1/tech/core-concepts/roadmap': '/docs/operate/roadmap/roadmap',
-    '/docs/v1/tech/guides/breached-password-detection': '/docs/operate/secure-and-monitor/breached-password-detection',
-    '/docs/v1/tech/reference/cors': '/docs/operate/secure-and-monitor/cors',
-    '/docs/v1/tech/core-concepts/key-master': '/docs/operate/secure-and-monitor/key-master',
-    '/docs/v1/tech/tutorials/key-rotation': '/docs/operate/secure-and-monitor/key-rotation',
-    '/docs/v1/tech/admin-guide/monitor': '/docs/operate/secure-and-monitor/monitor',
-    '/docs/v1/tech/tutorials/prometheus': '/docs/operate/secure-and-monitor/prometheus',
-    '/docs/v1/tech/admin-guide/securing': '/docs/operate/secure-and-monitor/securing',
-    '/docs/v1/tech/admin-guide/vulnerabilities': '/docs/operate/secure-and-monitor/vulnerabilities',
-    '/docs/v1/tech/admin-guide/technical-support': '/docs/operate/troubleshooting/technical-support',
-    '/docs/v1/tech/admin-guide/troubleshooting': '/docs/operate/troubleshooting/troubleshooting',
-    '/docs/v1/tech/getting-started/5-minute-docker': '/docs/quickstarts/5-minute-docker',
-    '/docs/v1/tech/getting-started/5-minute-fastpath': '/docs/quickstarts/5-minute-fastpath',
-    '/docs/v1/tech/5-minute-setup-guide': '/docs/quickstarts/5-minute-setup-guide',
-    '/docs/v1/tech/getting-started/5-minute-sandbox': '/docs/quickstarts/5-minute-sandbox',
-    '/docs/v1/tech/reference/configuration': '/docs/reference/configuration',
-    '/docs/v1/tech/reference/data-types': '/docs/reference/data-types',
-    '/docs/v1/tech/reference/cookies': '/docs/reference/cookies',
-    '/docs/v1/tech/reference/password-hashes': '/docs/reference/password-hashes',
-    '/docs/v1/tech/archive/release-notes': '/docs/archive/release-notes',
-    '/docs/v1/tech/release-notes': '/docs/release-notes/release-notes',
-    '/docs/v1/tech/example-apps/netcore': '/docs/sdks/examples/netcore',
-    '/docs/v1/tech/client-libraries/netcore': '/docs/sdks/netcore',
-    '/docs/v1/tech/client-libraries/angular-sdk': '/docs/sdks/angular-sdk',
-    '/docs/v1/tech/example-apps/dart': '/docs/sdks/examples/dart',
-    '/docs/v1/tech/client-libraries/dart': '/docs/sdks/dart',
-    '/docs/v1/tech/example-apps/go': '/docs/sdks/examples/go',
-    '/docs/v1/tech/client-libraries/go': '/docs/sdks/go',
-    '/docs/v1/tech/example-apps/java': '/docs/sdks/examples/java',
-    '/docs/v1/tech/client-libraries/java': '/docs/sdks/java',
-    '/docs/v1/tech/example-apps/javascript': '/docs/sdks/examples/javascript',
-    '/docs/v1/tech/client-libraries/javascript': '/docs/sdks/javascript',
-    '/docs/v1/tech/client-libraries/node': '/docs/sdks/node',
-    '/docs/v1/tech/client-libraries/openapi': '/docs/sdks/openapi',
-    '/docs/v1/tech/example-apps/php': '/docs/sdks/examples/php',
-    '/docs/v1/tech/client-libraries/php': '/docs/sdks/php',
-    '/docs/v1/tech/example-apps/python': '/docs/sdks/examples/python',
-    '/docs/v1/tech/client-libraries/python': '/docs/sdks/python',
-    '/docs/v1/tech/client-libraries/react-sdk': '/docs/sdks/react-sdk',
-    '/docs/v1/tech/example-apps/ruby': '/docs/sdks/examples/ruby',
-    '/docs/v1/tech/client-libraries/ruby': '/docs/sdks/ruby',
-    '/docs/v1/tech/client-libraries/typescript': '/docs/sdks/typescript',
-    '/docs/v1/tech/client-libraries/vue-sdk': '/docs/sdks/vue-sdk',
-    '/docs/v1/tech/account-management/': '/docs/lifecycle/manage-users/account-management/',
-    '/docs/v1/tech/apis/two-factor-1-25': '/docs/archive/apis/two-factor-1-25',
-    '/docs/v1/tech/advanced-threat-detection/': '/docs/operate/secure-and-monitor/advanced-threat-detection',
-    '/docs/v1/tech/archive/apis/system': '/docs/archive/apis/system-1-29',
-    '/docs/v1/tech/client-libraries/': '/docs/sdks/',
-    '/docs/v1/tech/events-webhooks/events/': '/docs/extend/events-and-webhooks/events/',
-    '/docs/v1/tech/getting-started/': '/docs/getting-started/index',
-    '/docs/v1/tech/identity-providers/': '/docs/lifecycle/authenticate-users/identity-providers/',
-    '/docs/v1/tech/identity-providers/samlv2-idp-initiated/': '/docs/lifecycle/authenticate-users/identity-providers/enterprise/samlv2-idp-initiated',
-    '/docs/v1/tech/installation-guide/kubernetes/': '/docs/get-started/download-and-install/kubernetes/index',
-    '/docs/v1/tech/installation-guide/migration': '/archive/passport-to-fusionauth',
-    '/docs/v1/tech/migration-guide/': '/docs/lifecycle/migrate-users/index',
-    '/docs/v1/tech/premium-features/': '/docs/get-started/core-concepts/premium-features',
-    '/docs/v1/tech/reference/': '/docs/reference/',
-    '/docs/v1/tech/': '/docs/',
-    '/docs/v1/tech/admin-guide/release-notifications': '/docs/operate/roadmap/releases',
-    '/docs/v1/tech/apis/consent': '/docs/apis/consents',
-    '/docs/v1/tech/common-errors': '/docs/operate/troubleshooting/troubleshooting',
-    '/docs/v1/tech/guides/migration': '/docs/lifecycle/migrate-users/bulk/general',
-    '/docs/v1/tech/guides/migration/general': '/docs/lifecycle/migrate-users/bulk/general',
-    '/docs/v1/tech/guides/passwordless': '/docs/lifecycle/authenticate-users/passwordless/magic-links',
-    '/docs/v1/tech/guides/webauthn': '/docs/lifecycle/authenticate-users/passwordless/webauthn-passkeys',
-    '/docs/v1/tech/installation-guide/configuration-management': '/docs/operate/deploy/configuration-management',
-    '/docs/v1/tech/installation-guide/monitor': '/docs/operate/secure-and-monitor/monitor',
-    '/docs/v1/tech/installation-guide/securing': '/docs/operate/secure-and-monitor/securing',
-    '/docs/v1/tech/installation-guide/upgrade': '/docs/operate/deploy/upgrade',
-    '/docs/v1/tech/passwordless/webauthn': '/docs/lifecycle/authenticate-users/passwordless/webauthn-passkeys',
-    '/docs/v1/tech/plugins/password-encryptors': '/docs/extend/code/password-hashes/custom-password-hashing',
-    '/docs/v1/tech/reference/password-encryptors': '/docs/reference/password-hashes',
-    '/docs/v1/tech/reactor': '/docs/get-started/core-concepts/licensing',
-    '/docs/v1/tech/reactor/': '/docs/get-started/core-concepts/licensing',
-    '/docs/v1/tech/troubleshooting': '/docs/operate/troubleshooting/troubleshooting',
-    '/docs/v1/tech/tutorials/gate-accounts-until-verified': '/docs/lifecycle/manage-users/verification/gate-accounts-until-user-email-verified',
-    '/docs/v1/tech/tutorials/migrate-users': '/docs/lifecycle/migrate-users/bulk/tutorial',
-    '/docs/v1/tech/tutorials/setting-up-user-account-lockout': '/docs/lifecycle/authenticate-users/setting-up-user-account-lockout',
-    '/docs/v1/tech/tutorials/two-factor/authenticator-app': '/docs/customize/email-and-messages/deprecated/authenticator-app-pre-1-26',
-    '/docs/v1/tech/tutorials/two-factor/twilio-push': '/docs/customize/email-and-messages/deprecated/twilio-push-pre-1-26',
-    '/docs/v1/tech/tutorials/integrate-angular': '/docs/quickstarts/quickstart-javascript-angular-web',
-    '/docs/v1/tech/tutorials/integrate-python-django': '/docs/quickstarts/quickstart-python-django-web',
-    '/docs/v1/tech/tutorials/integrate-python-flask': '/docs/quickstarts/quickstart-python-flask-web',
-    '/docs/v1/tech/tutorials/integrate-ruby-rails': '/docs/quickstarts/quickstart-ruby-rails-web',
-    '/docs/v1/tech/tutorials/integrate-dotnet': '/docs/quickstarts/quickstart-dotnet-web',
-    '/docs/v1/tech/tutorials/integrate-java-spring': '/docs/quickstarts/quickstart-springboot-web',
-    '/docs/v1/tech/tutorials/integrate-react': '/docs/quickstarts/quickstart-javascript-react-web',
-    '/learn/expert-advice/': '/articles/',
-    '/learn/expert-advice/authentication/gaming-identity-provider-needs': '/articles/gaming-entertainment/gaming-identity-provider-needs',
-    '/learn/expert-advice/authentication/login-authentication-workflows': '/articles/login-authentication-workflows/authentication-workflows-overview',
-    '/learn/expert-advice/authentication/login-authentication-workflows/': '/articles/login-authentication-workflows/authentication-workflows-overview',
-    '/learn/expert-advice/identity-basics/avoid-lockin': '/articles/authentication/avoid-lockin',
-    '/learn/expert-advice/identity-basics/common-authentication-implementation-risks': '/articles/authentication/common-authentication-implementation-risks',
-    '/learn/expert-advice/identity-basics/making-sure-your-auth-system-scales': '/articles/ciam/making-sure-your-auth-system-scales',
-    '/learn/expert-advice/identity-basics/value-standards-compliant-authentication': '/articles/oauth/value-standards-compliant-authentication',
-    '/learn/expert-advice/tokens/anatomy-of-jwt': '/articles/tokens/jwt-components-explained',
-    '/learn/expert-advice/ciam/ciam-vs-iam': '/articles/ciam/ciam-vs-iam',
-    '/learn/expert-advice/tokens/jwt-authentication-token-components-explained': '/articles/tokens/jwt-components-explained',
-    '/blog/archive/category/announcement': '/blog/category/news',
-    '/blog/archive/category/article': '/blog/category/education',
-    '/blog/archive/category/comparison': '/blog/category/compare',
-    '/blog/archive/category/community': '/blog/category/community',
-    '/blog/archive/category/features': '/blog/category/product',
-    '/blog/archive/category/tutorial': '/blog/category/tutorial',
-    '/dev-tools/jwt-debugger': '/dev-tools/jwt-decoder',
-    '/kubernetes/': '/docs/get-started/download-and-install/kubernetes/',
-    '/cognito': '/docs/lifecycle/migrate-users/bulk/cognito',
-    '/articles/oauth/what-is-oauth': '/articles/oauth/modern-guide-to-oauth',
-    '/podcast/': '/',
-    '/pricing/cloud': '/pricing',
-    '/pricing/edition': '/pricing',
-    '/pricing/edition/': '/pricing',
-    '/pricing/editions': '/pricing',
-    '/pricing/editions/': '/pricing',
-    '/products/identity-user-management/ciam-vs-iam': '/articles/ciam/ciam-vs-iam',
-    '/resources/guide-to-user-data-security': '/articles/security/guide-to-user-data-security',
-    '/learn/expert-advice/security/guide-to-user-data-security': '/articles/security/guide-to-user-data-security',
-    '/features/architecture': '/platform/built-for-developers',
-    '/features/built-for-developers': '/platform/built-for-developers',
-    '/features/advanced-registration-forms': '/platform/registration-forms',
-    '/features/breached-password-detection': '/features/authentication',
-    '/features/connectors': '/features/authentication',
-    '/features/scalability': '/platform/scalable',
-    '/features/security-data-compliance': '/security-data-compliance',
-    '/features/user-experience': '/platform/customizable',
-    '/features/user-management-reporting': '/features/user-management',
-    '/gaming': '/industries/gaming-entertainment',
-    '/gaming/': '/industries/gaming-entertainment',
-    '/podcast': '/',
-    '/pricing/cloud/': '/pricing',
-    '/pricing/edition': '/pricing',
-    '/pricing/edition/': '/pricing',
-    '/pricing/editions': '/pricing',
-    '/pricing/editions/': '/pricing',
-    '/resources/auth0-migration': '/auth0-migration',
-    '/upgrade/from-homegrown': '/compare',
-    '/upgrade/from-open-source': '/compare',
-    '/upgrade/from-saas': '/compare',
-    '/kubernetes': '/docs/get-started/download-and-install/kubernetes/index'
-  },
-
-  indexPaths: {
-    '/articles/': true,
-    '/articles/oauth/': true,
-    '/articles/security/': true,
-    '/articles/gaming-entertainment/': true,
-    '/articles/ciam/': true,
-    '/articles/login-authentication-workflows/': true,
-    '/articles/login-authentication-workflows/webapp/': true,
-    '/articles/login-authentication-workflows/spa/': true,
-    '/articles/login-authentication-workflows/mobile/': true,
-    '/articles/identity-basics/': true,
-    '/articles/authentication/': true,
-    '/articles/tokens/': true,
-    '/blog/': true,
-    '/dev-tools/': true,
-    '/docs/': true,
-    '/docs/extend/': true,
-    '/docs/extend/events-and-webhooks/': true,
-    '/docs/extend/events-and-webhooks/kafka/': true,
-    '/docs/extend/events-and-webhooks/events/': true,
-    '/docs/extend/code/': true,
-    '/docs/extend/code/lambdas/': true,
-    '/docs/extend/code/password-hashes/': true,
-    '/docs/extend/examples/': true,
-    '/docs/extend/examples/api-gateways/': true,
-    '/docs/apis/': true,
-    '/docs/apis/connectors/': true,
-    '/docs/apis/scim/': true,
-    '/docs/apis/custom-forms/': true,
-    '/docs/apis/messengers/': true,
-    '/docs/apis/identity-providers/': true,
-    '/docs/apis/entities/': true,
-    '/docs/sdks/': true,
-    '/docs/sdks/examples/': true,
-    '/docs/archive/': true,
-    '/docs/archive/apis/': true,
-    '/docs/lifecycle/': true,
-    '/docs/lifecycle/manage-users/': true,
-    '/docs/lifecycle/manage-users/account-management/': true,
-    '/docs/lifecycle/manage-users/search/': true,
-    '/docs/lifecycle/manage-users/verification/': true,
-    '/docs/lifecycle/authenticate-users/': true,
-    '/docs/lifecycle/authenticate-users/oauth/': true,
-    '/docs/lifecycle/authenticate-users/login-api/': true,
-    '/docs/lifecycle/authenticate-users/integrations/': true,
-    '/docs/lifecycle/authenticate-users/integrations/oidc/': true,
-    '/docs/lifecycle/authenticate-users/integrations/saml/': true,
-    '/docs/lifecycle/authenticate-users/identity-providers/': true,
-    '/docs/lifecycle/authenticate-users/identity-providers/gaming/': true,
-    '/docs/lifecycle/authenticate-users/identity-providers/enterprise/': true,
-    '/docs/lifecycle/authenticate-users/identity-providers/social/': true,
-    '/docs/lifecycle/authenticate-users/identity-providers/external-jwt/': true,
-    '/docs/lifecycle/authenticate-users/passwordless/': true,
-    '/docs/lifecycle/authenticate-users/saml/': true,
-    '/docs/lifecycle/migrate-users/': true,
-    '/docs/lifecycle/migrate-users/connectors/': true,
-    '/docs/lifecycle/migrate-users/scim/': true,
-    '/docs/lifecycle/migrate-users/bulk/': true,
-    '/docs/lifecycle/register-users/': true,
-    '/docs/release-notes/': true,
-    '/docs/operate/': true,
-    '/docs/operate/roadmap/': true,
-    '/docs/operate/deploy/': true,
-    '/docs/operate/troubleshooting/': true,
-    '/docs/operate/secure-and-monitor/': true,
-    '/docs/reference/': true,
-    '/docs/get-started/': true,
-    '/docs/get-started/download-and-install/': true,
-    '/docs/get-started/download-and-install/development/': true,
-    '/docs/get-started/download-and-install/kubernetes/': true,
-    '/docs/get-started/core-concepts/': true,
-    '/docs/get-started/run-in-the-cloud/': true,
-    '/docs/customize/': true,
-    '/docs/customize/email-and-messages/': true,
-    '/docs/customize/email-and-messages/deprecated/': true,
-    '/docs/customize/look-and-feel/': true,
-    '/docs/quickstarts/': true
-  },
-  redirectsByPrefix: [
-    ['/learn/expert-advice/dev-tools', '/dev-tools'],
-    ['/learn/expert-advice/authentication/spa', '/articles/login-authentication-workflows/spa'],
-    ['/learn/expert-advice/authentication/mobile', '/articles/login-authentication-workflows/mobile'],
-    ['/learn/expert-advice/authentication/webapp', '/articles/login-authentication-workflows/webapp'],
-    ['/learn/expert-advice', '/articles']
-  ],
-  redirectsByRegex: [
-    ['^/blog/(category|tag|author)/([^/]*)$', '$&/'],
-    ['/blog/archive/tag/', '/blog/tag/'],
-    ['/blog/\\d\\d\\d\\d/\\d\\d/\\d\\d/', '/blog/']
-  ],
-  s3Paths: ['/direct-download', '/license'],
-  s3Prefixes: ['/blog/', '/docs/', '/legal/', '/articles/', '/dev-tools/']
-}
-
-exports.handler = (event, context, callback) => {
-  var req = event.Records[0].cf.request;
-  var uri = req.uri;
-  var result = req;
-  
-  if (uri.endsWith('.html')) {
-
-    // there shouldn't be links to .html files out there
-    result = redir(uri.substring(0, uri.length - 5));
-
-  } else if (uri.endsWith('/') && removeSlash(uri)) {
-
-    // if this seems to be a reference to a file but with a trailing slash, remove it
-    result = redir(uri.substring(0, uri.length - 1));
-
-  } else if (!uri.endsWith('/') && referenceData.indexPaths[uri + '/'] === true) {
-
-    // if this is a reference to an index page that is missing its slash, add it
-    result = redir(uri + '/');
-  } else {
-
-    var redirect = calculateRedirect(uri);
-  
-    if (redirect !== null) {
-      result = redir(redirect);
-    } else {
-      req.uri = calculateURI(uri);
-    }
-  }
-  
-  callback(null, result);
-}
-
-// returns true if this seems to be a link to a page with a trailing slash that should just be removed
-function removeSlash(uri) {
-  return referenceData.indexPaths[uri] !== true &&
-         (!uri.startsWith('/blog') || (uri.match('^/blog/[\\w\\d-]*/$') && !uri.match('^/blog/latest/$'))) &&
-         referenceData.redirectsByPrefix.find(e => uri.startsWith(e[0])) === undefined;
-}
-
-function calculateRedirect(uri) {
-  var result = referenceData.redirects.hasOwnProperty(uri) ? referenceData.redirects[uri] : null;
-
-  if (result === null) {
-    var prefix_replacement = referenceData.redirectsByPrefix.find(e => uri.startsWith(e[0]));
-
-    if (prefix_replacement !== undefined) {
-      result = uri.replace(prefix_replacement[0], prefix_replacement[1]);
-    }
-  }
-
-  if (result === null) {
-    referenceData.redirectsByRegex.forEach(function (regexValueArray) {
-      var regex = new RegExp(regexValueArray[0], "g");
-      var value = regexValueArray[1];
-      if (regex.test(uri)) {
-        result = uri.replace(regex, value);
-      }
-    });
-  }
-
-  return result;
-}
-
-function appendHTML(uri) {
-  var slashIndex = uri.lastIndexOf('/');
-  var dotIndex = uri.indexOf('.', slashIndex);
-  if (slashIndex < uri.length - 1 && dotIndex < 0) {
-    uri = uri + '.html';
-  }
-  return uri;
-}
-
-function calculateURI(uri) {
-  var i;
-  for (i = 0; i < referenceData.s3Prefixes.length; i++) {
-    if (uri.startsWith(referenceData.s3Prefixes[i])) {
-      return uri.endsWith('/') ? uri + 'index.html' : appendHTML(uri);
-    }
-  }
-
-  for (i = 0; i < referenceData.s3Paths.length; i++) {
-    if (uri === referenceData.s3Paths[i]) {
-      return appendHTML(uri);
-    }
-  }
-
-  return uri;
-}
-
-function redir(loc) {
-  return {status:301,statusDescription:'Moved',headers:{'location':[{key: 'Location', value: loc}]}}
-}
